requirements.txt
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
GoogleAppEngineCloudStorageClient==1.9.22.1
